reconstrunct the folder setting
