--- lib/createDataKey.js
@@ -1,27 +1,0 @@
-const {distinguisher} = require('./config');
-const url = require('url');
-
-module.exports = function(req) {
-
-  // we use properties in distinguisher or url resource name
-  // to generate key for the data as their filename.
-  if(distinguisher.length == 0){
-    let key = getResourceNameByURL(req.url);
-    return {key: key, ok: true};
-  }else{
-    let key = '';
-    for(int i = 0; i < distinguisher.length; i++){
-      if(!req.hasOwnProperty(distinguisher[i])){
-        return {key: '', ok: false};
-      }else{
-        key += req.distinguisher[i];
-      }
-    }
-    return {key: key, ok: true};
-  }
-
-  function getResourceNameByURL(URL){
-    var nameArray = url.parse(URL).pathname.split('/');
-    return nameArray[nameArray.length-1];
-  }
-}
