+++ lib/file/fileAction.js
@@ -10,11 +10,15 @@ module.exports = {
     });
   }
 
-  readConfig: function read(){
+  readMockData: function readMockData(){
+    //  TODO:
+  }
+
+  readConfig: function readConfig(){
     // TODO:
   }
 
-  writeConfig: function writeConfig(){
+  createConfig: function createConfig(){
     // TODO:
   }
