+++ lib/server.js
@@ -1,10 +1,10 @@
 const {distinguisher, proxy_url, } = require('./config');
 const isReachable = require('is-reachable');
+const http = require('http');
+const httpProxy = require('http-proxy');
+const {mock_handler, record_handler} = require('./handlers')
 
 module.exports = function(option){
-  var http = require('http'),
-      httpProxy = require('http-proxy');
-
   var proxy = httpProxy.createProxyServer({target: proxy_url});
 
   /* we listen on the proxyReq event, Once we receive the request
@@ -22,6 +22,7 @@ module.exports = function(option){
         if(reachable){
           console.log('Good job, the target server is reachable!');
         }else {
+          console.log('ooops, can not reach target server, use mockdata instead' );
           mock_handler(proxyReq, res);
         }
     });
@@ -31,6 +32,7 @@ module.exports = function(option){
   /* if we can connect the target server
     we use record_handler to deal with it.
   */
+
 
   proxy.on('proxyRes', function(proxyRes, req, res){
     record_handler(proxyRes, req);
